+++ src/main/java/com/tsp/new_tsp_front/api/festival/domain/FrontFestivalEntity.java
@@ -20,7 +20,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 @SuperBuilder
 @EqualsAndHashCode(of = "idx", callSuper = false)
 @AllArgsConstructor
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "tsp_festival")
 public class FrontFestivalEntity extends NewCommonMappedClass {
