+++ src/main/java/com/tsp/new_tsp_front/api/model/domain/negotiation/FrontNegotiationEntity.java
@@ -25,7 +25,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 @SuperBuilder
 @EqualsAndHashCode(of = "idx", callSuper = false)
 @AllArgsConstructor
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "tsp_model_negotiation")
 public class FrontNegotiationEntity extends NewCommonMappedClass {
     @Transient
