+++ src/main/java/com/tsp/new_tsp_front/api/model/domain/schedule/FrontScheduleEntity.java
@@ -26,7 +26,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 @SuperBuilder
 @EqualsAndHashCode(of = "idx", callSuper = false)
 @AllArgsConstructor
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "tsp_model_schedule")
 public class FrontScheduleEntity extends NewCommonMappedClass {
     @Transient
