+++ src/test/java/com/tsp/new_tsp_front/api/festival/service/FrontFestivalJpaQueryRepositoryTest.java
@@ -13,6 +13,7 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.event.EventListener;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.TestPropertySource;
 
@@ -20,8 +21,6 @@ import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace.NONE;
@@ -35,11 +34,11 @@ import static org.springframework.test.context.TestConstructor.AutowireMode.ALL;
 @AutoConfigureTestDatabase(replace = NONE)
 @ExtendWith(MockitoExtension.class)
 @DisplayName("행사 Repository Test")
-class FrontFestivalJpaRepositoryTest {
+class FrontFestivalJpaQueryRepositoryTest {
 
     @Mock
-    private final FrontFestivalJpaRepository mockFrontFestivalJpaRepository;
-    private final FrontFestivalJpaRepository frontFestivalJpaRepository;
+    private final FrontFestivalJpaQueryRepository mockFrontFestivalJpaQueryRepository;
+    private final FrontFestivalJpaQueryRepository frontFestivalJpaQueryRepository;
 
     private FrontFestivalEntity frontFestivalEntity;
     private FrontFestivalDTO frontFestivalDTO;
@@ -73,7 +72,7 @@ class FrontFestivalJpaRepositoryTest {
     @DisplayName("행사 리스트 조회 테스트")
     void 행사리스트조회테스트() {
         // then
-        assertThat(frontFestivalJpaRepository.findFestivalList(frontFestivalEntity)).isNotEmpty();
+        assertThat(frontFestivalJpaQueryRepository.findFestivalList(frontFestivalEntity, PageRequest.of(1, 10))).isNotEmpty();
     }
 
     @Test
@@ -105,12 +104,6 @@ class FrontFestivalJpaRepositoryTest {
         em.flush();
         em.clear();
 
-        assertThat(frontFestivalJpaRepository.findFestivalGroup(dateTime.getMonthValue())).isNotEmpty();
-    }
-
-    @Test
-    @DisplayName("행사 상세 조회 테스트")
-    void 행사상세조회테스트() {
-        assertThat(frontFestivalJpaRepository.findOneFestival(frontFestivalDTO.getIdx()).getFestivalTitle()).isEqualTo("축제 제목");
+        assertThat(frontFestivalJpaQueryRepository.findFestivalGroup(dateTime.getMonthValue())).isNotEmpty();
     }
 }
